+++ src/basic/components/NotificationContainer.tsx
@@ -2,9 +2,9 @@ import { Notification } from '../../types';
 
 interface NotificationContainerProps {
   notifications: Notification[];
-  setNotifications: React.Dispatch<React.SetStateAction<Notification[]>>;
+  removeNotification: (id: string) => void;
 }
-const NotificationContainer = ({ notifications, setNotifications }: NotificationContainerProps) => {
+const NotificationContainer = ({ notifications, removeNotification }: NotificationContainerProps) => {
   if (notifications.length === 0) return null;
 
   return (
@@ -17,10 +17,7 @@ const NotificationContainer = ({ notifications, setNotifications }: Notification
           }`}
         >
           <span className="mr-2">{notif.message}</span>
-          <button
-            onClick={() => setNotifications(prev => prev.filter(n => n.id !== notif.id))}
-            className="text-white hover:text-gray-200"
-          >
+          <button onClick={() => removeNotification(notif.id)} className="text-white hover:text-gray-200">
             <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
               <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
             </svg>
